fix: user views folder name corrected
