+++ apps/platform/src/router/router.tsx
@@ -5,7 +5,7 @@ import { SignUpView } from '../views/auth/signUp.view'
 import { pages } from '@isomera/impl'
 import { PasswordResetView } from '../views/auth/passwordReset.view'
 import { PasswordResetConfirmView } from '../views/auth/passwordResetConfirm.view'
-import { UserInfoView } from '../views/user /info.view'
+import { UserInfoView } from '../views/user/info.view'
 import PublicRoute from './publicRoute'
 import { VerificationCodeView } from '../views/auth/verificationCode.view'
 
@@ -13,7 +13,7 @@ import { Routes, Route } from 'react-router-dom'
 import PrivateRoute from './privateRoute'
 import { PublicLayout } from '../layouts/public.layout'
 import { PrivateLayout } from '../layouts/private.layout'
-import { UserSecurityView } from '../views/user /profile-security.view'
+import { UserSecurityView } from '../views/user/profile-security.view'
 import { Verify2FAView } from '../views/auth/auth2FA.view'
 import { Recovery2FAView } from '../views/auth/recovery.view'
 import { Disable2FAView } from '../views/auth/disable-2fa.view'
